+++ source_gen/lib/builder.dart
@@ -32,12 +32,14 @@ Builder combiningBuilder([BuilderOptions options = BuilderOptions.empty]) {
   final ignoreForFile = Set<String>.from(
     optionsMap.remove('ignore_for_file') as List? ?? <String>[],
   );
+  final preamble = optionsMap.remove('preamble') as String? ?? '';
   final buildExtensions =
       validatedBuildExtensionsFrom(optionsMap, _defaultExtensions);
 
   final builder = CombiningBuilder(
     includePartName: includePartName,
     ignoreForFile: ignoreForFile,
+    preamble: preamble,
     buildExtensions: buildExtensions,
   );
 
@@ -58,6 +60,8 @@ class CombiningBuilder implements Builder {
 
   final Set<String> _ignoreForFile;
 
+  final String _preamble;
+
   @override
   final Map<String, List<String>> buildExtensions;
 
@@ -69,12 +73,14 @@ class CombiningBuilder implements Builder {
   const CombiningBuilder({
     bool? includePartName,
     Set<String>? ignoreForFile,
+    String? preamble,
     this.buildExtensions = _defaultExtensions,
   })  : _includePartName = includePartName ?? false,
-        _ignoreForFile = ignoreForFile ?? const <String>{};
+        _ignoreForFile = ignoreForFile ?? const <String>{},
+        _preamble = preamble ?? '';
 
   @override
-  Future build(BuildStep buildStep) async {
+  Future<void> build(BuildStep buildStep) async {
     // Pattern used for `findAssets`, which must be glob-compatible
     final pattern = buildStep.inputId.changeExtension('.*$_partFiles').path;
 
@@ -83,14 +89,16 @@ class CombiningBuilder implements Builder {
 
     // Pattern used to ensure items are only considered if they match
     // [file name without extension].[valid part id].[part file extension]
-    final restrictedPattern = RegExp([
-      '^', // start of string
-      RegExp.escape(inputBaseName), // file name, without extension
-      '.', // `.` character
-      partIdRegExpLiteral, // A valid part ID
-      RegExp.escape(_partFiles), // the ending part extension
-      '\$', // end of string
-    ].join());
+    final restrictedPattern = RegExp(
+      [
+        '^', // start of string
+        RegExp.escape(inputBaseName), // file name, without extension
+        r'\.', // `.` character
+        partIdRegExpLiteral, // A valid part ID
+        RegExp.escape(_partFiles), // the ending part extension
+        '\$', // end of string
+      ].join(),
+    );
 
     final assetIds = await buildStep
         .findAssets(Glob(pattern))
@@ -119,17 +127,22 @@ class CombiningBuilder implements Builder {
         await buildStep.resolver.compilationUnitFor(buildStep.inputId);
     final part = computePartUrl(buildStep.inputId, outputId);
     if (!hasExpectedPartDirective(libraryUnit, part)) {
-      log.warning('$part must be included as a part directive in '
-          'the input library with:\n    part \'$part\';');
+      log.warning(
+        '$part must be included as a part directive in '
+        'the input library with:\n    part \'$part\';',
+      );
       return;
     }
 
     final ignoreForFile = _ignoreForFile.isEmpty
         ? ''
         : '\n// ignore_for_file: ${_ignoreForFile.join(', ')}\n';
+
+    final preamble = _preamble.isEmpty ? '' : '\n$_preamble\n';
+
     final output = '''
 $defaultFileHeader
-${languageOverrideForLibrary(inputLibrary)}$ignoreForFile
+${languageOverrideForLibrary(inputLibrary)}$ignoreForFile$preamble
 part of $partOf;
 
 $assets