+++ source_gen/lib/src/builder.dart
@@ -2,6 +2,9 @@
 // for details. All rights reserved. Use of this source code is governed by a
 // BSD-style license that can be found in the LICENSE file.
 
+//TODO(kevmoo): https://github.com/dart-lang/linter/issues/3563
+// ignore_for_file: use_super_parameters
+
 import 'dart:convert';
 
 import 'package:analyzer/dart/ast/ast.dart';
@@ -60,29 +63,37 @@ class _Builder extends Builder {
         formatOutput = formatOutput ?? _formatter.format,
         _header = (header ?? defaultFileHeader).trim() {
     if (_generatedExtension.isEmpty || !_generatedExtension.startsWith('.')) {
-      throw ArgumentError.value(_generatedExtension, 'generatedExtension',
-          'Extension must be in the format of .*');
+      throw ArgumentError.value(
+        _generatedExtension,
+        'generatedExtension',
+        'Extension must be in the format of .*',
+      );
     }
     if (_isLibraryBuilder && _generators.length > 1) {
       throw ArgumentError(
-          'A standalone file can only be generated from a single Generator.');
+        'A standalone file can only be generated from a single Generator.',
+      );
     }
     if (options != null && additionalOutputExtensions.isNotEmpty) {
       throw ArgumentError(
-          'Either `options` or `additionalOutputExtensions` parameter '
-          'can be given. Not both.');
+        'Either `options` or `additionalOutputExtensions` parameter '
+        'can be given. Not both.',
+      );
     }
   }
 
   @override
-  Future build(BuildStep buildStep) async {
+  Future<void> build(BuildStep buildStep) async {
     final resolver = buildStep.resolver;
 
     if (!await resolver.isLibrary(buildStep.inputId)) return;
 
     if (_generators.every((g) => g is GeneratorForAnnotation) &&
         !(await _hasAnyTopLevelAnnotations(
-            buildStep.inputId, resolver, buildStep))) {
+          buildStep.inputId,
+          resolver,
+          buildStep,
+        ))) {
       return;
     }
 
@@ -91,7 +102,7 @@ class _Builder extends Builder {
     await _generateForLibrary(lib, buildStep);
   }
 
-  Future _generateForLibrary(
+  Future<void> _generateForLibrary(
     LibraryElement library,
     BuildStep buildStep,
   ) async {
@@ -128,8 +139,10 @@ class _Builder extends Builder {
             hasExpectedPartDirective(libraryUnit, part);
         if (!hasLibraryPartDirectiveWithOutputUri) {
           // TODO: Upgrade to error in a future breaking change?
-          log.warning('$part must be included as a part directive in '
-              'the input library with:\n    part \'$part\';');
+          log.warning(
+            '$part must be included as a part directive in '
+            'the input library with:\n    part \'$part\';',
+          );
           return;
         }
       } else {
@@ -221,10 +234,11 @@ class SharedPartBuilder extends _Builder {
         ) {
     if (!_partIdRegExp.hasMatch(partId)) {
       throw ArgumentError.value(
-          partId,
-          'partId',
-          '`partId` can only contain letters, numbers, `_` and `.`. '
-              'It cannot start or end with `.`.');
+        partId,
+        'partId',
+        '`partId` can only contain letters, numbers, `_` and `.`. '
+            'It cannot start or end with `.`.',
+      );
     }
   }
 }
@@ -354,7 +368,10 @@ Stream<GeneratedOutput> _generate(
 }
 
 Future<bool> _hasAnyTopLevelAnnotations(
-    AssetId input, Resolver resolver, BuildStep buildStep) async {
+  AssetId input,
+  Resolver resolver,
+  BuildStep buildStep,
+) async {
   if (!await buildStep.canRead(input)) return false;
   final parsed = await resolver.compilationUnitFor(input);
   final partIds = <AssetId>[];
@@ -377,7 +394,7 @@ Future<bool> _hasAnyTopLevelAnnotations(
   return false;
 }
 
-final _formatter = DartFormatter();
+final _formatter = DartFormatter(fixes: [StyleFix.singleCascadeStatements]);
 
 const defaultFileHeader = '// GENERATED CODE - DO NOT MODIFY BY HAND';