+++ source_gen/lib/src/constants/utils.dart
@@ -7,9 +7,9 @@ import 'package:analyzer/dart/element/element.dart';
 
 /// Throws a [FormatException] if [root] does not have a given field [name].
 ///
-/// Super types [ClassElement.supertype] are also checked before throwing.
-void assertHasField(ClassElement root, String name) {
-  ClassElement? element = root;
+/// Super types [InterfaceElement.supertype] are also checked before throwing.
+void assertHasField(InterfaceElement root, String name) {
+  InterfaceElement? element = root;
   while (element != null) {
     final field = element.getField(name);
     if (field != null) {
@@ -17,8 +17,11 @@ void assertHasField(ClassElement root, String name) {
     }
     element = element.supertype?.element;
   }
-  final allFields = root.fields.toSet()
-    ..addAll(root.allSupertypes.expand((t) => t.element.fields));
+  final allFields = {
+    ...root.fields,
+    for (var t in root.allSupertypes) ...t.element.fields,
+  };
+
   throw FormatException(
     'Class ${root.name} does not have field "$name".',
     'Fields: \n  - ${allFields.map((e) => e.name).join('\n  - ')}',