+++ source_gen/test/constants_test.dart
@@ -12,7 +12,8 @@ void main() {
     late List<ConstantReader> constants;
 
     setUpAll(() async {
-      final library = await resolveSource(r'''
+      final library = await resolveSource(
+        r'''
         library test_lib;
 
         const aString = 'Hello';
@@ -57,9 +58,11 @@ void main() {
         class Super extends Example {
           const Super() : super(aString: 'Super Hello');
         }
-      ''', (resolver) async => (await resolver.findLibraryByName('test_lib'))!);
+      ''',
+        (resolver) async => (await resolver.findLibraryByName('test_lib'))!,
+      );
       constants = library
-          .getType('Example')!
+          .getClass('Example')!
           .metadata
           .map((e) => ConstantReader(e.computeConstantValue()!))
           .toList();
@@ -117,17 +120,24 @@ void main() {
     test('should read a list', () {
       expect(constants[6].isList, isTrue, reason: '${constants[6]}');
       expect(constants[6].isLiteral, isTrue);
-      expect(constants[6].listValue.map((c) => ConstantReader(c).intValue),
-          [1, 2, 3]);
+      expect(
+        constants[6].listValue.map((c) => ConstantReader(c).intValue),
+        [1, 2, 3],
+      );
     });
 
     test('should read a map', () {
       expect(constants[7].isMap, isTrue, reason: '${constants[7]}');
       expect(constants[7].isLiteral, isTrue);
       expect(
-          constants[7].mapValue.map((k, v) => MapEntry(
-              ConstantReader(k!).intValue, ConstantReader(v!).stringValue)),
-          {1: 'A', 2: 'B'});
+        constants[7].mapValue.map(
+              (k, v) => MapEntry(
+                ConstantReader(k!).intValue,
+                ConstantReader(v!).stringValue,
+              ),
+            ),
+        {1: 'A', 2: 'B'},
+      );
     });
 
     test('should read a double', () {
@@ -195,13 +205,14 @@ void main() {
     });
   });
 
-  group('Reviable', () {
+  group('Revivable', () {
     late List<ConstantReader> constants;
 
     setUpAll(() async {
       final library = await resolveSource(
         r'''
         library test_lib;
+        import 'dart:io';
 
         @Int64Like.ZERO
         @Duration(seconds: 30)
@@ -216,6 +227,7 @@ void main() {
         @PublicWithPrivateConstructor._()
         @_privateField
         @Wrapper(_privateFunction)
+        @ProcessStartMode.normal
         class Example {}
 
         class Int64Like implements Int64LikeBase{
@@ -289,7 +301,7 @@ void main() {
         (resolver) async => (await resolver.findLibraryByName('test_lib'))!,
       );
       constants = library
-          .getType('Example')!
+          .getClass('Example')!
           .metadata
           .map((e) => ConstantReader(e.computeConstantValue()))
           .toList();
@@ -373,6 +385,14 @@ void main() {
       expect(function.isPrivate, isTrue);
       expect(function.source.fragment, isEmpty);
       expect(function.accessor, '_privateFunction');
+    });
+
+    test('should decode public static fields backed by private constructors',
+        () {
+      final staticFieldWithPrivateImpl = constants[13].revive();
+      expect(staticFieldWithPrivateImpl.accessor, 'ProcessStartMode.normal');
+      expect(staticFieldWithPrivateImpl.isPrivate, isFalse);
+      expect(staticFieldWithPrivateImpl.source.fragment, isEmpty);
     });
   });
 }