+++ source_gen/test/external_only_type_checker_test.dart
@@ -7,6 +7,8 @@
 
 // Increase timeouts on this test which resolves source code and can be slow.
 @Timeout.factor(2.0)
+library test;
+
 import 'package:analyzer/dart/element/nullability_suffix.dart';
 import 'package:analyzer/dart/element/type.dart';
 import 'package:build/build.dart';
@@ -21,15 +23,24 @@ void main() {
 
   setUpAll(() async {
     late LibraryReader thisTest;
-    await resolveSource(r'''
+    await resolveSource(
+      r'''
       export 'type_checker_test.dart' show NonPublic;
-    ''', (resolver) async {
-      thisTest = LibraryReader(await resolver.libraryFor(
-          AssetId('source_gen', 'test/external_only_type_checker_test.dart')));
-    }, inputId: AssetId('source_gen', 'test/example.dart'));
+    ''',
+      (resolver) async {
+        thisTest = LibraryReader(
+          await resolver.libraryFor(
+            AssetId('source_gen', 'test/external_only_type_checker_test.dart'),
+          ),
+        );
+      },
+      inputId: AssetId('source_gen', 'test/example.dart'),
+    );
 
     staticNonPublic = thisTest.findType('NonPublic')!.instantiate(
-        typeArguments: const [], nullabilitySuffix: NullabilitySuffix.none);
+      typeArguments: const [],
+      nullabilitySuffix: NullabilitySuffix.none,
+    );
     staticNonPublicChecker = TypeChecker.fromStatic(staticNonPublic);
   });
 
@@ -39,25 +50,35 @@ void main() {
   }) {
     group('NonPublic', () {
       test('should equal NonPublic', () {
-        expect(checkNonPublic().isExactlyType(staticNonPublic), isTrue,
-            reason: '${checkNonPublic()} != ${staticNonPublic.element.name}');
+        expect(
+          checkNonPublic().isExactlyType(staticNonPublic),
+          isTrue,
+          reason: '${checkNonPublic()} != ${staticNonPublic.element.name}',
+        );
       });
 
       test('should be assignable from NonPublic', () {
-        expect(checkNonPublic().isAssignableFromType(staticNonPublic), isTrue,
-            reason: '${checkNonPublic()} is not assignable from '
-                '${staticNonPublic.element.name}');
+        expect(
+          checkNonPublic().isAssignableFromType(staticNonPublic),
+          isTrue,
+          reason: '${checkNonPublic()} is not assignable from '
+              '${staticNonPublic.element.name}',
+        );
       });
     });
   }
 
-  group('TypeChecker.forRuntime', () {
-    commonTests(
-      checkNonPublic: () => const TypeChecker.fromRuntime(NonPublic),
-    );
-  }, onPlatform: const {
-    'windows': Skip('https://github.com/dart-lang/source_gen/issues/573')
-  });
+  group(
+    'TypeChecker.forRuntime',
+    () {
+      commonTests(
+        checkNonPublic: () => const TypeChecker.fromRuntime(NonPublic),
+      );
+    },
+    onPlatform: const {
+      'windows': Skip('https://github.com/dart-lang/source_gen/issues/573')
+    },
+  );
 
   group('TypeChecker.forStatic', () {
     commonTests(
@@ -68,7 +89,8 @@ void main() {
   group('TypeChecker.fromUrl', () {
     commonTests(
       checkNonPublic: () => const TypeChecker.fromUrl(
-          'asset:source_gen/test/external_only_type_checker_test.dart#NonPublic'),
+        'asset:source_gen/test/external_only_type_checker_test.dart#NonPublic',
+      ),
     );
   });
 }