+++ source_gen/test/generator_for_annotation_test.dart
@@ -4,6 +4,8 @@
 
 // The first test that runs `testBuilder` takes a LOT longer than the rest.
 @Timeout.factor(3)
+library test;
+
 import 'package:analyzer/dart/analysis/utilities.dart';
 import 'package:analyzer/dart/ast/ast.dart';
 import 'package:analyzer/dart/element/element.dart';
@@ -18,7 +20,7 @@ void main() {
       '`null`': null,
       'empty string': '',
       'only whitespace': '\n \t',
-      'empty list': [],
+      'empty list': <Object>[],
       'list with null, empty, and whitespace items': [null, '', '\n \t']
     }.entries) {
       test(entry.key, () async {
@@ -36,8 +38,11 @@ void main() {
       yield '// ${element.name}';
     });
     final builder = LibraryBuilder(generator);
-    await testBuilder(builder, _inputMap, outputs: {
-      'a|lib/file.g.dart': r'''
+    await testBuilder(
+      builder,
+      _inputMap,
+      outputs: {
+        'a|lib/file.g.dart': r'''
 // GENERATED CODE - DO NOT MODIFY BY HAND
 
 // **************************************************************************
@@ -52,7 +57,8 @@ void main() {
 
 // baz
 '''
-    });
+      },
+    );
   });
 
   group('handles errors correctly', () {
@@ -85,30 +91,42 @@ void main() {
   test('Does not resolve the library if there are no top level annotations',
       () async {
     final builder =
-        LibraryBuilder(_StubGenerator<Deprecated>('Deprecated', (_) {}));
+        LibraryBuilder(_StubGenerator<Deprecated>('Deprecated', (_) => null));
     final input = AssetId('a', 'lib/a.dart');
     final assets = {input: 'main() {}'};
 
     final reader = InMemoryAssetReader(sourceAssets: assets);
     final resolver = _TestingResolver(assets);
 
-    await runBuilder(builder, [input], reader, InMemoryAssetWriter(),
-        _FixedResolvers(resolver));
+    await runBuilder(
+      builder,
+      [input],
+      reader,
+      InMemoryAssetWriter(),
+      _FixedResolvers(resolver),
+    );
 
     expect(resolver.parsedUnits, {input});
     expect(resolver.resolvedLibs, isEmpty);
   });
 
   test('applies to annotated libraries', () async {
-    final builder = LibraryBuilder(_StubGenerator<Deprecated>(
-        'Deprecated', (element) => '// ${element.displayName}'));
-    await testBuilder(builder, {
-      'a|lib/file.dart': '''
+    final builder = LibraryBuilder(
+      _StubGenerator<Deprecated>(
+        'Deprecated',
+        (element) => '// ${element.displayName}',
+      ),
+    );
+    await testBuilder(
+      builder,
+      {
+        'a|lib/file.dart': '''
       @deprecated
       library foo;
       '''
-    }, outputs: {
-      'a|lib/file.g.dart': '''
+      },
+      outputs: {
+        'a|lib/file.g.dart': '''
 // GENERATED CODE - DO NOT MODIFY BY HAND
 
 // **************************************************************************
@@ -117,7 +135,8 @@ void main() {
 
 // foo
 '''
-    });
+      },
+    );
   });
 }
 
@@ -129,7 +148,10 @@ class _StubGenerator<T> extends GeneratorForAnnotation<T> {
 
   @override
   Object? generateForAnnotatedElement(
-          Element element, ConstantReader annotation, BuildStep buildStep) =>
+    Element element,
+    ConstantReader annotation,
+    BuildStep buildStep,
+  ) =>
       _behavior(element);
 
   @override
@@ -157,8 +179,10 @@ class _TestingResolver implements ReleasableResolver {
   _TestingResolver(this.assets);
 
   @override
-  Future<CompilationUnit> compilationUnitFor(AssetId assetId,
-      {bool allowSyntaxErrors = false}) async {
+  Future<CompilationUnit> compilationUnitFor(
+    AssetId assetId, {
+    bool allowSyntaxErrors = false,
+  }) async {
     parsedUnits.add(assetId);
     return parseString(content: assets[assetId]!).unit;
   }
@@ -170,10 +194,12 @@ class _TestingResolver implements ReleasableResolver {
   }
 
   @override
-  Future<LibraryElement> libraryFor(AssetId assetId,
-      {bool allowSyntaxErrors = false}) async {
+  Future<LibraryElement> libraryFor(
+    AssetId assetId, {
+    bool allowSyntaxErrors = false,
+  }) async {
     resolvedLibs.add(assetId);
-    return null as LibraryElement;
+    throw StateError('This method intentionally throws');
   }
 
   @override