+++ source_gen/test/type_checker_test.dart
@@ -2,8 +2,12 @@
 // for details. All rights reserved. Use of this source code is governed by a
 // BSD-style license that can be found in the LICENSE file.
 
+// ignore_for_file: unreachable_from_main
+
 // Increase timeouts on this test which resolves source code and can be slow.
 @Timeout.factor(2.0)
+library test;
+
 import 'dart:collection';
 
 import 'package:analyzer/dart/element/element.dart';
@@ -20,9 +24,13 @@ void main() {
   late InterfaceType staticMap;
   late InterfaceType staticHashMap;
   late InterfaceType staticUnmodifiableListView;
+  late InterfaceType staticEnum;
   late TypeChecker staticIterableChecker;
+  late TypeChecker staticEnumMixinChecker;
   late TypeChecker staticMapChecker;
+  late TypeChecker staticMapMixinChecker;
   late TypeChecker staticHashMapChecker;
+  late TypeChecker staticEnumChecker;
 
   // Resolved top-level types from package:source_gen.
   late InterfaceType staticGenerator;
@@ -30,33 +38,85 @@ void main() {
   late TypeChecker staticGeneratorChecker;
   late TypeChecker staticGeneratorForAnnotationChecker;
 
+  // Resolved top-level types from this file
+  late InterfaceType staticEnumMixin;
+  late InterfaceType staticMapMixin;
+  late InterfaceType staticMyEnum;
+  late InterfaceType staticMyEnumWithMixin;
+
   setUpAll(() async {
     late LibraryElement core;
     late LibraryElement collection;
     late LibraryReader sourceGen;
-    await resolveSource(r'''
+    late LibraryElement testSource;
+    await resolveSource(
+      r'''
       export 'package:source_gen/source_gen.dart';
       export 'type_checker_test.dart' show NonPublic;
-    ''', (resolver) async {
-      core = (await resolver.findLibraryByName('dart.core'))!;
-      collection = (await resolver.findLibraryByName('dart.collection'))!;
-      sourceGen = LibraryReader(await resolver
-          .libraryFor(AssetId('source_gen', 'lib/source_gen.dart')));
-    }, inputId: AssetId('source_gen', 'test/example.dart'));
+    ''',
+      (resolver) async {
+        core = (await resolver.findLibraryByName('dart.core'))!;
+        collection = (await resolver.findLibraryByName('dart.collection'))!;
+        sourceGen = LibraryReader(
+          await resolver
+              .libraryFor(AssetId('source_gen', 'lib/source_gen.dart')),
+        );
+        testSource = await resolver
+            .libraryFor(AssetId('source_gen', 'test/type_checker_test.dart'));
+      },
+      inputId: AssetId('source_gen', 'test/example.dart'),
+    );
 
-    final staticIterable = core.getType('Iterable')!.instantiate(
-        typeArguments: [core.typeProvider.dynamicType],
-        nullabilitySuffix: NullabilitySuffix.none);
+    final staticIterable = core.getClass('Iterable')!.instantiate(
+      typeArguments: [core.typeProvider.dynamicType],
+      nullabilitySuffix: NullabilitySuffix.none,
+    );
     staticIterableChecker = TypeChecker.fromStatic(staticIterable);
-    staticUri = core.getType('Uri')!.instantiate(
-        typeArguments: [], nullabilitySuffix: NullabilitySuffix.none);
-    staticMap = core.getType('Map')!.instantiate(typeArguments: [
-      core.typeProvider.dynamicType,
-      core.typeProvider.dynamicType
-    ], nullabilitySuffix: NullabilitySuffix.none);
+    staticUri = core.getClass('Uri')!.instantiate(
+      typeArguments: [],
+      nullabilitySuffix: NullabilitySuffix.none,
+    );
+    staticMap = core.getClass('Map')!.instantiate(
+      typeArguments: [
+        core.typeProvider.dynamicType,
+        core.typeProvider.dynamicType
+      ],
+      nullabilitySuffix: NullabilitySuffix.none,
+    );
     staticMapChecker = TypeChecker.fromStatic(staticMap);
+    staticEnum = core.getClass('Enum')!.instantiate(
+      typeArguments: [],
+      nullabilitySuffix: NullabilitySuffix.none,
+    );
+    staticEnumChecker = TypeChecker.fromStatic(staticEnum);
+    staticEnumMixin =
+        (testSource.exportNamespace.get('MyEnumMixin')! as InterfaceElement)
+            .instantiate(
+      typeArguments: [],
+      nullabilitySuffix: NullabilitySuffix.none,
+    );
+    staticEnumMixinChecker = TypeChecker.fromStatic(staticEnumMixin);
+    staticMapMixin =
+        (testSource.exportNamespace.get('MyMapMixin')! as InterfaceElement)
+            .instantiate(
+      typeArguments: [],
+      nullabilitySuffix: NullabilitySuffix.none,
+    );
+    staticMapMixinChecker = TypeChecker.fromStatic(staticMapMixin);
+    staticMyEnum =
+        (testSource.exportNamespace.get('MyEnum')! as InterfaceElement)
+            .instantiate(
+      typeArguments: [],
+      nullabilitySuffix: NullabilitySuffix.none,
+    );
+    staticMyEnumWithMixin =
+        (testSource.exportNamespace.get('MyEnumWithMixin')! as InterfaceElement)
+            .instantiate(
+      typeArguments: [],
+      nullabilitySuffix: NullabilitySuffix.none,
+    );
 
-    staticHashMap = collection.getType('HashMap')!.instantiate(
+    staticHashMap = collection.getClass('HashMap')!.instantiate(
       typeArguments: [
         core.typeProvider.dynamicType,
         core.typeProvider.dynamicType
@@ -64,20 +124,22 @@ void main() {
       nullabilitySuffix: NullabilitySuffix.none,
     );
     staticHashMapChecker = TypeChecker.fromStatic(staticHashMap);
-    staticUnmodifiableListView = collection
-        .getType('UnmodifiableListView')!
-        .instantiate(
-            typeArguments: [core.typeProvider.dynamicType],
-            nullabilitySuffix: NullabilitySuffix.none);
+    staticUnmodifiableListView =
+        collection.getClass('UnmodifiableListView')!.instantiate(
+      typeArguments: [core.typeProvider.dynamicType],
+      nullabilitySuffix: NullabilitySuffix.none,
+    );
 
     staticGenerator = sourceGen.findType('Generator')!.instantiate(
-        typeArguments: [], nullabilitySuffix: NullabilitySuffix.none);
+      typeArguments: [],
+      nullabilitySuffix: NullabilitySuffix.none,
+    );
     staticGeneratorChecker = TypeChecker.fromStatic(staticGenerator);
-    staticGeneratorForAnnotation = sourceGen
-        .findType('GeneratorForAnnotation')!
-        .instantiate(
-            typeArguments: [core.typeProvider.dynamicType],
-            nullabilitySuffix: NullabilitySuffix.none);
+    staticGeneratorForAnnotation =
+        sourceGen.findType('GeneratorForAnnotation')!.instantiate(
+      typeArguments: [core.typeProvider.dynamicType],
+      nullabilitySuffix: NullabilitySuffix.none,
+    );
     staticGeneratorForAnnotationChecker =
         TypeChecker.fromStatic(staticGeneratorForAnnotation);
   });
@@ -85,7 +147,10 @@ void main() {
   // Run a common set of type comparison checks with various implementations.
   void commonTests({
     required TypeChecker Function() checkIterable,
+    required TypeChecker Function() checkEnum,
     required TypeChecker Function() checkMap,
+    required TypeChecker Function() checkMapMixin,
+    required TypeChecker Function() checkEnumMixin,
     required TypeChecker Function() checkHashMap,
     required TypeChecker Function() checkGenerator,
     required TypeChecker Function() checkGeneratorForAnnotation,
@@ -93,15 +158,52 @@ void main() {
     group('(Iterable)', () {
       test('should be assignable from dart:collection#UnmodifiableListView',
           () {
-        expect(checkIterable().isAssignableFromType(staticUnmodifiableListView),
-            true);
+        expect(
+          checkIterable().isAssignableFromType(staticUnmodifiableListView),
+          true,
+        );
       });
     });
 
+    group('(Enum)', () {
+      test('should be supertype of enum classes', () {
+        expect(checkEnum().isSuperTypeOf(staticMyEnum), isTrue);
+      });
+
+      test(
+        'with mixins should be assignable to mixin class',
+        () {
+          expect(
+            checkEnumMixin().isAssignableFromType(staticMyEnumWithMixin),
+            isTrue,
+          );
+        },
+        onPlatform: const {
+          'windows': Skip('https://github.com/dart-lang/source_gen/issues/573'),
+        },
+      );
+    });
+
+    group(
+      '(MapMixin',
+      () {
+        test('should equal MapMixin class', () {
+          expect(checkMapMixin().isExactlyType(staticMapMixin), isTrue);
+          expect(checkMapMixin().isExactly(staticMapMixin.element), isTrue);
+        });
+      },
+      onPlatform: const {
+        'windows': Skip('https://github.com/dart-lang/source_gen/issues/573'),
+      },
+    );
+
     group('(Map)', () {
       test('should equal dart:core#Map', () {
-        expect(checkMap().isExactlyType(staticMap), isTrue,
-            reason: '${checkMap()} != ${staticMap.element.name}');
+        expect(
+          checkMap().isExactlyType(staticMap),
+          isTrue,
+          reason: '${checkMap()} != ${staticMap.element.name}',
+        );
       });
 
       test('should not be a super type of dart:core#Map', () {
@@ -109,8 +211,11 @@ void main() {
       });
 
       test('should not equal dart:core#HashMap', () {
-        expect(checkMap().isExactlyType(staticHashMap), isFalse,
-            reason: '${checkMap()} == $staticHashMapChecker');
+        expect(
+          checkMap().isExactlyType(staticHashMap),
+          isFalse,
+          reason: '${checkMap()} == $staticHashMapChecker',
+        );
       });
 
       test('should be a super type of dart:collection#HashMap', () {
@@ -119,6 +224,10 @@ void main() {
 
       test('should be assignable from dart:collection#HashMap', () {
         expect(checkMap().isAssignableFromType(staticHashMap), isTrue);
+      });
+
+      test('should be assignable from dart:collection#MapMixin', () {
+        expect(checkMap().isAssignableFromType(staticMapMixin), isTrue);
       });
 
       // Ensure we're consistent WRT generic types
@@ -134,8 +243,11 @@ void main() {
 
     group('(HashMap)', () {
       test('should equal dart:collection#HashMap', () {
-        expect(checkHashMap().isExactlyType(staticHashMap), isTrue,
-            reason: '${checkHashMap()} != $staticHashMapChecker');
+        expect(
+          checkHashMap().isExactlyType(staticHashMap),
+          isTrue,
+          reason: '${checkHashMap()} != $staticHashMapChecker',
+        );
       });
 
       test('should not be a super type of dart:core#Map', () {
@@ -149,29 +261,38 @@ void main() {
 
     group('(Generator)', () {
       test('should equal Generator', () {
-        expect(checkGenerator().isExactlyType(staticGenerator), isTrue,
-            reason: '${checkGenerator()} != ${staticGenerator.element.name}');
+        expect(
+          checkGenerator().isExactlyType(staticGenerator),
+          isTrue,
+          reason: '${checkGenerator()} != ${staticGenerator.element.name}',
+        );
       });
 
       test('should not be a super type of Generator', () {
-        expect(checkGenerator().isSuperTypeOf(staticGenerator), isFalse,
-            reason: '${checkGenerator()} is super of '
-                '${staticGenerator.element.name}');
+        expect(
+          checkGenerator().isSuperTypeOf(staticGenerator),
+          isFalse,
+          reason: '${checkGenerator()} is super of '
+              '${staticGenerator.element.name}',
+        );
       });
 
       test('should be a super type of GeneratorForAnnotation', () {
-        expect(checkGenerator().isSuperTypeOf(staticGeneratorForAnnotation),
-            isTrue,
-            reason: '${checkGenerator()} is not super of '
-                '${staticGeneratorForAnnotation.element.name}');
+        expect(
+          checkGenerator().isSuperTypeOf(staticGeneratorForAnnotation),
+          isTrue,
+          reason: '${checkGenerator()} is not super of '
+              '${staticGeneratorForAnnotation.element.name}',
+        );
       });
 
       test('should be assignable from GeneratorForAnnotation', () {
         expect(
-            checkGenerator().isAssignableFromType(staticGeneratorForAnnotation),
-            isTrue,
-            reason: '${checkGenerator()} is not assignable from '
-                '${staticGeneratorForAnnotation.element.name}');
+          checkGenerator().isAssignableFromType(staticGeneratorForAnnotation),
+          isTrue,
+          reason: '${checkGenerator()} is not assignable from '
+              '${staticGeneratorForAnnotation.element.name}',
+        );
       });
     });
   }
@@ -179,7 +300,10 @@ void main() {
   group('TypeChecker.forRuntime', () {
     commonTests(
       checkIterable: () => const TypeChecker.fromRuntime(Iterable),
+      checkEnum: () => const TypeChecker.fromRuntime(Enum),
+      checkEnumMixin: () => const TypeChecker.fromRuntime(MyEnumMixin),
       checkMap: () => const TypeChecker.fromRuntime(Map),
+      checkMapMixin: () => const TypeChecker.fromRuntime(MyMapMixin),
       checkHashMap: () => const TypeChecker.fromRuntime(HashMap),
       checkGenerator: () => const TypeChecker.fromRuntime(Generator),
       checkGeneratorForAnnotation: () =>
@@ -190,7 +314,10 @@ void main() {
   group('TypeChecker.forStatic', () {
     commonTests(
       checkIterable: () => staticIterableChecker,
+      checkEnum: () => staticEnumChecker,
+      checkEnumMixin: () => staticEnumMixinChecker,
       checkMap: () => staticMapChecker,
+      checkMapMixin: () => staticMapMixinChecker,
       checkHashMap: () => staticHashMapChecker,
       checkGenerator: () => staticGeneratorChecker,
       checkGeneratorForAnnotation: () => staticGeneratorForAnnotationChecker,
@@ -200,12 +327,21 @@ void main() {
   group('TypeChecker.fromUrl', () {
     commonTests(
       checkIterable: () => const TypeChecker.fromUrl('dart:core#Iterable'),
+      checkEnum: () => const TypeChecker.fromUrl('dart:core#Enum'),
+      checkEnumMixin: () => const TypeChecker.fromUrl(
+        'asset:source_gen/test/type_checker_test.dart#MyEnumMixin',
+      ),
       checkMap: () => const TypeChecker.fromUrl('dart:core#Map'),
+      checkMapMixin: () => const TypeChecker.fromUrl(
+        'asset:source_gen/test/type_checker_test.dart#MyMapMixin',
+      ),
       checkHashMap: () => const TypeChecker.fromUrl('dart:collection#HashMap'),
       checkGenerator: () => const TypeChecker.fromUrl(
-          'package:source_gen/src/generator.dart#Generator'),
+        'package:source_gen/src/generator.dart#Generator',
+      ),
       checkGeneratorForAnnotation: () => const TypeChecker.fromUrl(
-          'package:source_gen/src/generator_for_annotation.dart#GeneratorForAnnotation'),
+        'package:source_gen/src/generator_for_annotation.dart#GeneratorForAnnotation',
+      ),
     );
   });
 
@@ -219,7 +355,7 @@ void main() {
     ''',
       (resolver) async => (await resolver.findLibraryByName('_test'))!,
     );
-    final classX = library.getType('X')!;
+    final classX = library.getClass('X')!;
     const $deprecated = TypeChecker.fromRuntime(Deprecated);
 
     expect(
@@ -290,16 +426,28 @@ void main() {
         (resolver) async => (await resolver.findLibraryByName('_test'))!,
       );
 
-      $A = TypeChecker.fromStatic(library.getType('A')!.instantiate(
-          typeArguments: [], nullabilitySuffix: NullabilitySuffix.none));
-      $B = TypeChecker.fromStatic(library.getType('B')!.instantiate(
-          typeArguments: [], nullabilitySuffix: NullabilitySuffix.none));
-      $C = TypeChecker.fromStatic(library.getType('C')!.instantiate(
-          typeArguments: [], nullabilitySuffix: NullabilitySuffix.none));
-      $ExampleOfA = library.getType('ExampleOfA')!;
-      $ExampleOfMultiA = library.getType('ExampleOfMultiA')!;
-      $ExampleOfAPlusB = library.getType('ExampleOfAPlusB')!;
-      $ExampleOfBPlusC = library.getType('ExampleOfBPlusC')!;
+      $A = TypeChecker.fromStatic(
+        library.getClass('A')!.instantiate(
+          typeArguments: [],
+          nullabilitySuffix: NullabilitySuffix.none,
+        ),
+      );
+      $B = TypeChecker.fromStatic(
+        library.getClass('B')!.instantiate(
+          typeArguments: [],
+          nullabilitySuffix: NullabilitySuffix.none,
+        ),
+      );
+      $C = TypeChecker.fromStatic(
+        library.getClass('C')!.instantiate(
+          typeArguments: [],
+          nullabilitySuffix: NullabilitySuffix.none,
+        ),
+      );
+      $ExampleOfA = library.getClass('ExampleOfA')!;
+      $ExampleOfMultiA = library.getClass('ExampleOfMultiA')!;
+      $ExampleOfAPlusB = library.getClass('ExampleOfAPlusB')!;
+      $ExampleOfBPlusC = library.getClass('ExampleOfBPlusC')!;
     });
 
     test('of a single @A', () {
@@ -359,9 +507,13 @@ void main() {
     ''',
         (resolver) async => (await resolver.findLibraryByName('_test'))!,
       );
-      $A = TypeChecker.fromStatic(library.getType('A')!.instantiate(
-          typeArguments: [], nullabilitySuffix: NullabilitySuffix.none));
-      $ExampleOfA = library.getType('ExampleOfA')!;
+      $A = TypeChecker.fromStatic(
+        library.getClass('A')!.instantiate(
+          typeArguments: [],
+          nullabilitySuffix: NullabilitySuffix.none,
+        ),
+      );
+      $ExampleOfA = library.getClass('ExampleOfA')!;
       $annotatedParameter = library.topLevelElements
           .whereType<FunctionElement>()
           .firstWhere((f) => f.name == 'annotatedParameter')
@@ -370,53 +522,86 @@ void main() {
     });
 
     test('should throw by default', () {
-      expect(() => $A.firstAnnotationOf($ExampleOfA),
-          throwsUnresolvedAnnotationException);
-      expect(() => $A.annotationsOf($ExampleOfA),
-          throwsUnresolvedAnnotationException);
-      expect(() => $A.firstAnnotationOfExact($ExampleOfA),
-          throwsUnresolvedAnnotationException);
-      expect(() => $A.annotationsOfExact($ExampleOfA),
-          throwsUnresolvedAnnotationException);
-      expect(() => $A.firstAnnotationOf($annotatedParameter),
-          throwsUnresolvedAnnotationException);
-      expect(() => $A.annotationsOf($annotatedParameter),
-          throwsUnresolvedAnnotationException);
-      expect(() => $A.firstAnnotationOfExact($annotatedParameter),
-          throwsUnresolvedAnnotationException);
-      expect(() => $A.annotationsOfExact($annotatedParameter),
-          throwsUnresolvedAnnotationException);
+      expect(
+        () => $A.firstAnnotationOf($ExampleOfA),
+        throwsUnresolvedAnnotationException,
+      );
+      expect(
+        () => $A.annotationsOf($ExampleOfA),
+        throwsUnresolvedAnnotationException,
+      );
+      expect(
+        () => $A.firstAnnotationOfExact($ExampleOfA),
+        throwsUnresolvedAnnotationException,
+      );
+      expect(
+        () => $A.annotationsOfExact($ExampleOfA),
+        throwsUnresolvedAnnotationException,
+      );
+      expect(
+        () => $A.firstAnnotationOf($annotatedParameter),
+        throwsUnresolvedAnnotationException,
+      );
+      expect(
+        () => $A.annotationsOf($annotatedParameter),
+        throwsUnresolvedAnnotationException,
+      );
+      expect(
+        () => $A.firstAnnotationOfExact($annotatedParameter),
+        throwsUnresolvedAnnotationException,
+      );
+      expect(
+        () => $A.annotationsOfExact($annotatedParameter),
+        throwsUnresolvedAnnotationException,
+      );
     });
 
     test('should not throw if `throwOnUnresolved` == false', () {
       expect(
-          $A
-              .firstAnnotationOf($ExampleOfA, throwOnUnresolved: false)!
-              .type!
-              .element!
-              .name,
-          'A');
+        $A
+            .firstAnnotationOf($ExampleOfA, throwOnUnresolved: false)!
+            .type!
+            .element!
+            .name,
+        'A',
+      );
       expect(
-          $A
-              .annotationsOf($ExampleOfA, throwOnUnresolved: false)
-              .map((a) => a.type!.element!.name),
-          ['A']);
+        $A
+            .annotationsOf($ExampleOfA, throwOnUnresolved: false)
+            .map((a) => a.type!.element!.name),
+        ['A'],
+      );
       expect(
-          $A
-              .firstAnnotationOfExact($ExampleOfA, throwOnUnresolved: false)!
-              .type!
-              .element!
-              .name,
-          'A');
+        $A
+            .firstAnnotationOfExact($ExampleOfA, throwOnUnresolved: false)!
+            .type!
+            .element!
+            .name,
+        'A',
+      );
       expect(
-          $A
-              .annotationsOfExact($ExampleOfA, throwOnUnresolved: false)
-              .map((a) => a.type!.element!.name),
-          ['A']);
+        $A
+            .annotationsOfExact($ExampleOfA, throwOnUnresolved: false)
+            .map((a) => a.type!.element!.name),
+        ['A'],
+      );
     });
   });
 }
 
 final throwsUnresolvedAnnotationException = throwsA(
-    const TypeMatcher<UnresolvedAnnotationException>()
-        .having((e) => e.annotationSource, 'annotationSource', isNotNull));
+  const TypeMatcher<UnresolvedAnnotationException>()
+      .having((e) => e.annotationSource, 'annotationSource', isNotNull),
+);
+
+// Not using `dart.collection#MapMixin` because we want to test elements
+// actually declared as a `mixin`.
+mixin MyMapMixin on Map<dynamic, dynamic> {}
+
+enum MyEnum { foo, bar }
+
+mixin MyEnumMixin on Enum {
+  int get value => 1;
+}
+
+enum MyEnumWithMixin with MyEnumMixin { foo, bar }